Support transparent server restart after upgrading
The build script now bakes the server version hash
number in proxy.php that it is built against,
such that the latter can use it to detect stale servers.

After an upgrade, proxy.php will be running the newer
version but, unless loolwsd is restarted, the server
version returned in hosting/capabilities will be
unexpected (i.e. will not match the one in proxy.php).
When a mismatch is detected, proxy.php will kill
the loolwsd process and start a new instance.

Also replaced a number of tabs with spaces and fixed
indentation issues to help readability and consistency.