Check for fontconfig in proxy.php?status.
fontconfig is necessary for the AppImage to work, but many server
systems do not have it.

I've considered adding it to the AppImage, but then there are many
reports of AppImages not working due to containing fontconfig older than
the system one in case the newer fontconfig contains changes in the
configuration files (like new xml tags) that are not known for the older
one.